Test for App_Config\Include directory path in Production.ps1 as well

+++ windows/9.2.0/sitecore-assets/tools/entrypoints/iis/Production.ps1
@@ -78,8 +78,10 @@ while ($true)
 # wait for application pool to stop
 Wait-WebItemState -IISPath "IIS:\AppPools\DefaultAppPool" -State "Stopped"
 
-# inject Sitecore config files
-Copy-Item -Path (Join-Path $PSScriptRoot "\*.config") -Destination "C:\inetpub\wwwroot\App_Config\Include"
+if (Test-Path -Path "C:\inetput\wwwroot\App_Config\Include") {
+    # inject Sitecore config files
+    Copy-Item -Path (Join-Path $PSScriptRoot "\*.config") -Destination "C:\inetpub\wwwroot\App_Config\Include"
+}
 
 # start ServiceMonitor.exe in background, kill foreground process if it fails
 Start-Job -Name "ServiceMonitor.exe" {
